refactor(cli): remove inquirer from cli-core
