+++ packages/@cdktf/cli-core/package.json
@@ -108,7 +108,6 @@
     "@types/fs-extra": "^8.1.2",
     "@types/ink": "^2.0.3",
     "@types/ink-spinner": "^3.0.1",
-    "@types/inquirer": "^8.2.5",
     "@types/jest": "^29.5.0",
     "@types/json-schema": "^7.0.11",
     "@types/lodash.isequal": "^4.5.6",
@@ -144,7 +143,6 @@
     "ink-spinner": "^4.0.3",
     "ink-testing-library": "^2.1.0",
     "ink-use-stdout-dimensions": "^1.0.5",
-    "inquirer": "^8.2.5",
     "jest": "^29.5.0",
     "jsii-srcmak": "^0.1.867",
     "lodash.isequal": "^4.5.0",
