+++ packages/cdktf-cli/src/bin/cmds/handlers.ts
@@ -61,6 +61,7 @@ import {
   verifySimilarLibraryVersion,
 } from "./helper/check-environment";
 import { sanitizeVarFiles } from "./helper/var-files";
+import { askForCrashReportingConsent } from "./helper/error-reporting";
 
 const chalkColour = new chalk.Instance();
 const config = readConfigSync();
@@ -118,7 +119,7 @@ export async function convert({ language, provider, stack }: any) {
 }
 
 export async function deploy(argv: any) {
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   throwIfNotProjectDirectory();
   await displayVersionMessage();
   await checkEnvironment();
@@ -167,7 +168,7 @@ export async function deploy(argv: any) {
 }
 
 export async function destroy(argv: any) {
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   throwIfNotProjectDirectory();
   await displayVersionMessage();
   await checkEnvironment();
@@ -202,7 +203,7 @@ export async function destroy(argv: any) {
 }
 
 export async function diff(argv: any) {
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   throwIfNotProjectDirectory();
   await displayVersionMessage();
   await checkEnvironment();
@@ -239,7 +240,7 @@ export async function get(argv: {
 }) {
   throwIfNotProjectDirectory();
   await displayVersionMessage();
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   await checkEnvironment();
   await verifySimilarLibraryVersion();
   const config = readConfigSync(); // read config again to be up-to-date (if called via 'add' command)
@@ -298,7 +299,7 @@ export async function init(argv: any) {
 }
 
 export async function list(argv: any) {
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   throwIfNotProjectDirectory();
   await displayVersionMessage();
   await checkEnvironment();
@@ -354,7 +355,7 @@ export async function login(argv: { tfeHostname: string }) {
 }
 
 export async function synth(argv: any) {
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   throwIfNotProjectDirectory();
   await displayVersionMessage();
   await checkEnvironment();
@@ -378,7 +379,7 @@ export async function synth(argv: any) {
 }
 
 export async function watch(argv: any) {
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   throwIfNotProjectDirectory();
   await displayVersionMessage();
   const command = argv.app;
@@ -408,7 +409,7 @@ export async function watch(argv: any) {
 }
 
 export async function output(argv: any) {
-  await initializErrorReporting(true);
+  await initializErrorReporting(askForCrashReportingConsent);
   throwIfNotProjectDirectory();
   await displayVersionMessage();
   await checkEnvironment();
